Update the Nova\Cache

+++ src/Cache/FileStore.php
@@ -8,7 +8,7 @@ use Nova\Filesystem\Filesystem;
 class FileStore implements StoreInterface
 {
     /**
-     * The Illuminate Filesystem instance.
+     * The Nova Filesystem instance.
      *
      * @var \Nova\Filesystem\Filesystem
      */
@@ -43,24 +43,50 @@ class FileStore implements StoreInterface
      */
     public function get($key)
     {
+        return array_get($this->getPayload($key), 'data');
+    }
+
+    /**
+     * Retrieve an item and expiry time from the cache by key.
+     *
+     * @param  string  $key
+     * @return array
+     */
+    protected function getPayload($key)
+    {
         $path = $this->path($key);
 
-        if (! $this->files->exists($path)) {
-            return null;
+        // If the file doesn't exists, we obviously can't return the cache so we will
+        // just return null. Otherwise, we'll get the contents of the file and get
+        // the expiration UNIX timestamps from the start of the file's contents.
+        if ( ! $this->files->exists($path)) {
+            return array('data' => null, 'time' => null);
         }
 
         try
         {
             $expire = substr($contents = $this->files->get($path), 0, 10);
         } catch (\Exception $e) {
-            return null;
+            return array('data' => null, 'time' => null);
         }
 
-        if (time() < $expire) {
-            return unserialize(substr($contents, 10));
+        // If the current time is greater than expiration timestamps we will delete
+        // the file and return null. This helps clean up the old files and keeps
+        // this directory much cleaner for us as old files aren't hanging out.
+        if (time() >= $expire) {
+            $this->forget($key);
+
+            return array('data' => null, 'time' => null);
         }
 
-        return $this->forget($key);
+        $data = unserialize(substr($contents, 10));
+
+        // Next, we'll extract the number of minutes that are remaining for a cache
+        // so that we can properly retain the time for things like the increment
+        // operation that may be performed on the cache. We'll round this out.
+        $time = ceil(($expire - time()) / 60);
+
+        return compact('data', 'time');
     }
 
     /**
@@ -101,13 +127,17 @@ class FileStore implements StoreInterface
      *
      * @param  string  $key
      * @param  mixed   $value
-     * @return void
-     *
-     * @throws \LogicException
+     * @return int
      */
     public function increment($key, $value = 1)
     {
-        throw new \LogicException("Increment operations not supported by this driver.");
+        $raw = $this->getPayload($key);
+
+        $int = ((int) $raw['data']) + $value;
+
+        $this->put($key, $int, (int) $raw['time']);
+
+        return $int;
     }
 
     /**
@@ -115,13 +145,11 @@ class FileStore implements StoreInterface
      *
      * @param  string  $key
      * @param  mixed   $value
-     * @return void
-     *
-     * @throws \LogicException
+     * @return int
      */
     public function decrement($key, $value = 1)
     {
-        throw new \LogicException("Decrement operations not supported by this driver.");
+        return $this->increment($key, $value * -1);
     }
 
     /**
@@ -146,7 +174,8 @@ class FileStore implements StoreInterface
     {
         $file = $this->path($key);
 
-        if ($this->files->exists($file)) {
+        if ($this->files->exists($file))
+        {
             $this->files->delete($file);
         }
     }
@@ -158,8 +187,10 @@ class FileStore implements StoreInterface
      */
     public function flush()
     {
-        foreach ($this->files->directories($this->directory) as $directory) {
-            $this->files->deleteDirectory($directory);
+        if ($this->files->isDirectory($this->directory)) {
+            foreach ($this->files->directories($this->directory) as $directory) {
+                $this->files->deleteDirectory($directory);
+            }
         }
     }
 
@@ -173,7 +204,7 @@ class FileStore implements StoreInterface
     {
         $parts = array_slice(str_split($hash = md5($key), 2), 0, 2);
 
-        return $this->directory.'/'.join('/', $parts).'/'.$hash;
+        return $this->directory .DS .join(DS, $parts) .DS .$hash;
     }
 
     /**
